# Fix the benchmark script to call `to_a`